refactor: removed vue router
