--- src/router/index.js
@@ -1,15 +1,0 @@
-import { createRouter, createWebHistory } from 'vue-router'
-import AppView from '../views/AppView.vue'
-
-const router = createRouter({
-  history: createWebHistory(import.meta.env.BASE_URL),
-  routes: [
-    {
-      path: '/',
-      name: 'app',
-      component: AppView
-    }
-  ]
-})
-
-export default router
